requirements.txt
amqp
billiard
brotlipy
cached-property
celery
certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
cryptography
Deprecated
Django
django-celery-beat
django-filter
django-oauth-toolkit
django-timezone-field
djangorestframework
future
greenlet
idna
importlib-metadata
jwcrypto
kombu
Markdown
oauthlib
persiantools
prompt-toolkit
pycparser
pyOpenSSL
PySocks
python-crontab
python-dateutil
pytz
redis
requests
schedule
six
SQLAlchemy
sqlparse
typing-extensions
urllib3
vine
wcwidth
win-inet-pton
wincertstore
wrapt
zipp
